Add a UI for selecting layout algorithm

+++ src/main/java/org/BioLayoutExpress3D/Utils/FloatNumberField.java
@@ -25,7 +25,7 @@ public final class FloatNumberField extends JTextField
 
     private NumberFormat numberFormatter = null;
 
-    public FloatNumberField(int value, int columns)
+    public FloatNumberField(float value, int columns)
     {
         super(columns);
